Remove monkey patching from the init file
